refactor: isolate display modes
