+++ packages/skott/bin/display-mode.ts
@@ -1,0 +1,14 @@
+import type EventEmitter from "events";
+
+export class DisplayMode {
+  constructor(
+    private readonly identifier: string,
+    private readonly displayFunction: () => void
+  ) {}
+
+  registerWatchMode(watchEmitter: EventEmitter) {
+    watchEmitter.on("change", () => {
+      this.displayFunction();
+    });
+  }
+}
